实现 Wiki 文档向量化与使用量计数
后端：
- Wiki 文档全流程：预上传、内容抽取、分区、元数据生成、向量化、下载、重命名、删除，新增 EmbeddingDocumentConsumer 与 pgvector 向量存储
- 新增 AI 模型/插件/Wiki 使用量计数器及后台聚合任务，扩展 WorkerTask 实体与 WorkerState
- 新增知识图谱（Neo4j）设置项与服务
- 移除旧的 WikiDocumentChunkEmbedding 实体，Wiki 配置与实体同步更新

前端：
- Wiki 页面重构为列表/详情/文档详情，新增团队 Wiki 页面
- Kiota 客户端迁移 Wiki 文档相关接口，同步 i18n 与路由

其他：
- 补充 Wiki/使用量计数器单元测试与 E2E 脚本
- 更新 docs、ui/docs 与 rounds-log

+++ src/aiplugin/MoAI.AIPlugin.Shared/Queries/Responses/PluginBaseInfoItem.cs
@@ -62,5 +62,5 @@ public class PluginBaseInfoItem : AuditsInfo
     /// <summary>
     /// 使用量计数.
     /// </summary>
-    public int Counter { get; set; }
+    public long Counter { get; set; }
 }
