+++ tool/PostgresScaffold/Database/Entities/WikiDocumentChunkMetadatumEntity.cs
@@ -47,7 +47,7 @@ public partial class WikiDocumentChunkMetadatumEntity : IFullAudited
     /// <summary>
     /// 创建人.
     /// </summary>
-    public int CreateUserId { get; set; }
+    public long CreateUserId { get; set; }
 
     /// <summary>
     /// 创建时间.
@@ -57,7 +57,7 @@ public partial class WikiDocumentChunkMetadatumEntity : IFullAudited
     /// <summary>
     /// 更新人.
     /// </summary>
-    public int UpdateUserId { get; set; }
+    public long UpdateUserId { get; set; }
 
     /// <summary>
     /// 更新时间.
